create PseudoRandom macro, remove unused s16 casts

+++ src/countdown.c
@@ -180,7 +180,7 @@ void sub_8036398(void)
     element->x = countdown->unk60 - gCamera.x;
     element->y = countdown->unk64 - gCamera.y;
 
-    if (IS_OUT_OF_CAM_RANGE(element->x, (s16)element->y)) {
+    if (IS_OUT_OF_CAM_RANGE(element->x, element->y)) {
         TaskDestroy(gCurTask);
         return;
     }
